tsx改为ts
