+++ src/config/index.ts
@@ -1,0 +1,1 @@
+export const serviceUrl: string = "http://localhost:3333/api";
