+++ src/utils/tools.ts
@@ -1,0 +1,9 @@
+import type { RcFile } from "antd/es/upload/interface";
+export const getBase64 = (file: RcFile): Promise<string> =>
+	new Promise((resolve, reject) => {
+		const reader = new FileReader();
+
+		reader.readAsDataURL(file);
+		reader.onload = () => resolve(reader.result as string);
+		reader.onerror = (error) => reject(error);
+	});
